🚚 api: certification-attestations-for-division (third route):
move controller to certification course src
