+++ api/lib/application/organizations/organization-controller.js
@@ -24,7 +24,6 @@ import dayjs from 'dayjs';
 import customParseFormat from 'dayjs/plugin/customParseFormat.js';
 dayjs.extend(customParseFormat);
 import { getDivisionCertificationResultsCsv } from '../../infrastructure/utils/csv/certification-results/get-division-certification-results-csv.js';
-import { getCertificationAttestationsPdf as certificationAttestationPdf } from '../../infrastructure/utils/pdf/certification-attestation-pdf.js';
 import * as organizationForAdminSerializer from '../../infrastructure/serializers/jsonapi/organizations-administration/organization-for-admin-serializer.js';
 
 import * as csvSerializer from '../../infrastructure/serializers/csv/csv-serializer.js';
@@ -188,35 +187,6 @@ const createOrganizationPlacesLot = async function (
   return h.response(dependencies.organizationPlacesLotManagementSerializer.serialize(organizationPlacesLot)).code(201);
 };
 
-const downloadCertificationAttestationsForDivision = async function (
-  request,
-  h,
-  dependencies = { certificationAttestationPdf },
-) {
-  const organizationId = request.params.id;
-  const { i18n } = request;
-  const { division, isFrenchDomainExtension } = request.query;
-
-  const attestations = await usecases.findCertificationAttestationsForDivision({
-    organizationId,
-    division,
-  });
-
-  const { buffer } = await dependencies.certificationAttestationPdf.getCertificationAttestationsPdfBuffer({
-    certificates: attestations,
-    isFrenchDomainExtension,
-    i18n,
-  });
-
-  const now = dayjs();
-  const fileName = `${now.format('YYYYMMDD')}_attestations_${division}.pdf`;
-
-  return h
-    .response(buffer)
-    .header('Content-Disposition', `attachment; filename=${fileName}`)
-    .header('Content-Type', 'application/pdf');
-};
-
 const downloadCertificationResults = async function (
   request,
   h,
@@ -375,7 +345,6 @@ const organizationController = {
   findOrganizationPlacesLot,
   deleteOrganizationPlacesLot,
   createOrganizationPlacesLot,
-  downloadCertificationAttestationsForDivision,
   downloadCertificationResults,
   findTargetProfiles,
   attachTargetProfiles,
