+++ api/src/certification/course/application/certification-attestation-controller.js
@@ -1,5 +1,6 @@
 import { usecases } from '../../shared/domain/usecases/index.js';
 import * as certificationAttestationPdf from '../infrastructure/utils/pdf/certification-attestation-pdf.js';
+import dayjs from 'dayjs';
 
 const getPDFAttestation = async function (request, h, dependencies = { certificationAttestationPdf }) {
   const userId = request.auth.credentials.userId;
@@ -49,8 +50,38 @@ const getCertificationPDFAttestationsForSession = async function (
     .header('Content-Type', 'application/pdf');
 };
 
+const downloadCertificationAttestationsForDivision = async function (
+  request,
+  h,
+  dependencies = { certificationAttestationPdf },
+) {
+  const organizationId = request.params.id;
+  const { i18n } = request;
+  const { division, isFrenchDomainExtension } = request.query;
+
+  const attestations = await usecases.findCertificationAttestationsForDivision({
+    organizationId,
+    division,
+  });
+
+  const { buffer } = await dependencies.certificationAttestationPdf.getCertificationAttestationsPdfBuffer({
+    certificates: attestations,
+    isFrenchDomainExtension,
+    i18n,
+  });
+
+  const now = dayjs();
+  const fileName = `${now.format('YYYYMMDD')}_attestations_${division}.pdf`;
+
+  return h
+    .response(buffer)
+    .header('Content-Disposition', `attachment; filename=${fileName}`)
+    .header('Content-Type', 'application/pdf');
+};
+
 const certificationAttestationController = {
   getPDFAttestation,
   getCertificationPDFAttestationsForSession,
+  downloadCertificationAttestationsForDivision,
 };
 export { certificationAttestationController };
